# docs/openapi/tkcreazy-seedance.openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TKCreazy Seedance 视频生成 API",
    "version": "1.4.2",
...
    "license": {
      "name": "LGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/lgpl-3.0.html"
    }
  },
  "servers": [
    {
      "url": "https://tkcreazy.top",
      "description": "TKCreazy 生产环境"
    }
  ],
...
  "security": [
    {
      "BearerAuth": []
    },
    {
      "XApiKeyAuth": []
    },
    {
      "XGoogApiKeyAuth": []
    }
  ],
  "paths": {
    "/api/v3/contents/generations/uploads": {
      "post": {
        "tags": [
          "Seedance 视频生成"
        ],
        "summary": "上传 Seedance 参考图片",
        "description": "### 推荐流程：先上传，再创建视频\n\n此接口只保存图片，不会创建视频。调用顺序如下：\n\n1. 使用准备创建视频的同一个 API Key 调用本接口。multipart 文件字段必须叫 image；也可以使用 JSON 的 image_base64。\n2. 成功响应同时返回 upload_id 和 image_url。下一步必须复制完整 image_url；upload_id 只用于标识记录，不能单独作为生成参数。\n3. 调用 POST /api/v3/contents/generations/tasks，在 content 数组中增加 image_url 项：image_url.url 填上一步返回的完整 URL。\n4. 如果图片是首帧，省略 role 或写 role=first_frame；如果图片只是视觉参考，写 role=reference_image；多张参考图需要分别上传，然后在 content 中重复多个 image_url 项。\n\nmultipart 文件上传示例：\n\n~~~bash\ncurl -X POST \"https://tkcreazy.top/api/v3/contents/generations/uploads\" \\\n  -H \"Authorization: Bearer <YOUR_API_KEY>\" \\\n  -F \"image=@reference.png;type=image/png\"\n~~~\n\n如果原始图片已经是 Base64，也可以把步骤 1 改为 JSON 上传：\n\n~~~json\n{\n  \"image_base64\": \"<完整 Base64，不带 data: 前缀>\",\n  \"content_type\": \"image/png\",\n  \"filename\": \"reference.png\"\n}\n~~~\n\n两种上传方式的响应结构相同。后续都只使用响应中的完整 image_url。\n\n步骤 1 的典型响应：\n\n~~~json\n{\n  \"upload_id\": \"sdupl_example_uploaded_image\",\n  \"image_url\": \"https://tkcreazy.top/api/v3/contents/generations/uploads/sdupl_example_uploaded_image\",\n  \"content_type\": \"image/png\",\n  \"size\": 72396,\n  \"sha256\": \"d42c2780ffcedeb699ac245f42118b4ed082e61d1ba7146549b9df08607f2731\",\n  \"expires_at\": \"2026-07-25T02:52:36Z\"\n}\n~~~\n\n上面的值是格式示例。真实请求必须使用自己当次响应里的 image_url。\n\n把参考图片保存到本站为 Seedance 单独使用的对象存储命名空间，并返回可在创建任务中引用的 image_url。请求可二选一：multipart/form-data 的文件字段必须命名为 image；application/json 使用 image_base64，可传裸 Base64（此时 content_type 必填）或完整 Data URI（MIME 已写在 URI 中，content_type 可省略）。只接受真实内容可解码的 PNG、JPEG、WebP，文件或 Base64 解码后最大 10 MiB，宽高各不超过 8192 像素且总像素不超过 4000 万；声明 MIME 与实际内容不一致会被拒绝。filename 仅作为可选客户端文件名元数据，不决定 COS 对象名或上游行为。\n\n响应 image_url 是本站受鉴权保护的托管图片地址，有效期见 expires_at；只能由上传时同一用户、同一 API Key、同一 Seedance 分组读取或用于创建任务。创建任务时网关识别该地址，从 COS 生成短期签名 URL 后转发给上游，所以上游不需要持有本站 API Key。上传图片不产生视频按秒费用。媒体存储未配置时返回 503 media_storage_not_configured；同一用户的全部 Seedance API Key 共用独立媒体 I/O 并发上限，达到限制时返回 429 media_concurrency_exceeded。expires_at 表示托管记录的可访问期限，不代表 Redis 到期会自动删除 COS 对象；管理员应为实际存储前缀下的 seedance/inputs/ 配置至少 2 天的 Bucket lifecycle，并清理 noncurrent versions。\n\n2026-07-24 生产实测：同一张 72,396 字节 PNG 分别以 multipart/form-data、JSON 裸 Base64 + content_type、JSON 完整 Data URI 提交，三次均返回 HTTP 200、image/png、size=72396 和相同 SHA-256。下面的 upload_id、URL 和时间仍是可导入 Apifox 的格式示例，实际调用必须使用当次响应值。",
        "operationId": "uploadSeedanceImage",
        "requestBody": {
          "required": true,
          "description": "这是两步流程的步骤 1。选择 multipart/form-data 或 application/json 其中一种上传方式；不要两种同时传。此接口不能上传音频或视频，也不会直接创建 Seedance 任务。",
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SeedanceMultipartImageUploadRequest"
              },
              "encoding": {
                "image": {
                  "contentType": "image/png, image/jpeg, image/webp"
                }
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeedanceBase64ImageUploadRequest"
              },
              "examples": {
                "bareBase64": {
                  "summary": "2026-07-24 生产实测成功：裸 Base64，content_type 必填",
                  "value": {
                    "image_base64": "iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8/x8AAusB9Y9Z5YQAAAAASUVORK5CYII=",
                    "content_type": "image/png",
                    "filename": "first-frame.png"
                  },
                  "description": "实测使用 72,396 字节 PNG，返回 HTTP 200，响应 size、MIME 与 SHA-256 均与原图一致。value 中的短 Base64 仅用于展示字段格式。"
                },
                "dataURI": {
                  "summary": "2026-07-24 生产实测成功：完整 Data URI，content_type 可省略",
                  "value": {
                    "image_base64": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAQAAAC1HAwCAAAAC0lEQVR42mP8/x8AAusB9Y9Z5YQAAAAASUVORK5CYII=",
                    "filename": "reference.png"
                  },
                  "description": "实测使用 data:image/png;base64,...，返回 HTTP 200，响应 size、MIME 与 SHA-256 均与原图一致。value 中的短 Data URI 仅用于展示字段格式。"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "步骤 1 成功：图片已写入 Seedance 媒体存储。下一步复制响应中的完整 image_url 到创建任务请求的 content[].image_url.url；不要只复制 upload_id。upload_id 和 image_url 只属于当前用户、API Key 与 Seedance 分组。",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeedanceImageUploadResponse"
                },
                "example": {
                  "upload_id": "sdupl_example_uploaded_image",
                  "image_url": "https://tkcreazy.top/api/v3/contents/generations/uploads/sdupl_example_uploaded_image",
                  "content_type": "image/png",
                  "size": 72396,
                  "sha256": "d42c2780ffcedeb699ac245f42118b4ed082e61d1ba7146549b9df08607f2731",
                  "expires_at": "2026-07-25T02:52:36Z"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "413": {
            "$ref": "#/components/responses/ImageTooLarge"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedUploadMediaType"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "default": {
            "$ref": "#/components/responses/UpstreamOrUnexpectedError"
          }
        }
      }
    },
    "/api/v3/contents/generations/uploads/{upload_id}": {
      "get": {
        "tags": [
          "Seedance 视频生成"
        ],
        "summary": "读取已上传的 Seedance 图片",
        "description": "从 Seedance 独立 COS/S3 对象前缀读取已上传图片。必须使用上传该图片时的同一用户、同一 API Key 和同一 Seedance 分组；地址不是公开图片链接。Redis 记录默认保留 24 小时，以 POST 上传响应中的 expires_at 为准；过期、已删除或所有权不匹配都返回 404 upload_not_found。Redis 记录到期不等于 COS 对象已删除，长期清理由管理员为 seedance/inputs/ 配置的 Bucket lifecycle 负责。支持单范围 Range 请求，实际 200/206/416 及响应头取决于对象存储；媒体 I/O 并发达到限制时返回 429。\n\n普通客户端通常无需主动调用本接口：把 POST 上传返回的完整 image_url 原样放入创建任务即可，网关会在服务端解析并签名 COS 对象。2026-07-24 生产实测：完整读取返回 HTTP 200、image/png、72,396 字节，SHA-256 与上传原图一致；Range bytes=0-1023 返回 HTTP 206、1024 字节、Accept-Ranges: bytes 和 Content-Range: bytes 0-1023/72396。",
        "operationId": "getSeedanceUploadedImage",
        "parameters": [
          {
            "$ref": "#/components/parameters/UploadId"
          },
          {
            "$ref": "#/components/parameters/MediaRange"
          }
        ],
        "responses": {
          "200": {
            "description": "完整图片。Content-Type 是实际检测到的 image/png、image/jpeg 或 image/webp。",
            "headers": {
              "Accept-Ranges": {
                "$ref": "#/components/headers/AcceptRanges"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLength"
              },
              "Content-Disposition": {
                "$ref": "#/components/headers/ContentDisposition"
              }
            },
            "content": {
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "206": {
            "description": "对象存储接受 Range 后返回的部分图片内容。",
            "headers": {
              "Accept-Ranges": {
                "$ref": "#/components/headers/AcceptRanges"
              },
              "Content-Length": {
                "$ref": "#/components/headers/ContentLength"
              },
              "Content-Range": {
                "$ref": "#/components/headers/ContentRange"
              }
            },
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/UploadNotFound"
          },
          "416": {
            "description": "请求范围不可满足。Content-Range 可能包含完整对象大小；错误体由对象存储决定，客户端应以 HTTP 状态为准。",
            "headers": {
              "Content-Range": {
                "$ref": "#/components/headers/ContentRange"
              }
            }
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          },
          "default": {
            "$ref": "#/components/responses/UpstreamOrUnexpectedError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "本站 API Key 的推荐鉴权方式。请求头格式：Authorization: Bearer YOUR_TKCREAZY_API_KEY。BearerAuth、XApiKeyAuth、XGoogApiKeyAuth 三者任选其一，不要同时配置，也不要在查询参数中传 key 或 api_key。"
      },
      "XApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "本站 API Key 的兼容鉴权方式。BearerAuth、XApiKeyAuth、XGoogApiKeyAuth 三者任选其一。"
      },
      "XGoogApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-goog-api-key",
        "description": "为 Gemini CLI 兼容保留的本站 API Key 请求头。BearerAuth、XApiKeyAuth、XGoogApiKeyAuth 三者任选其一。"
      }
    },
...
    "schemas": {
      "SeedanceMultipartImageUploadRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "image"
        ],
        "properties": {
          "image": {
            "type": "string",
            "format": "binary",
            "description": "必须使用字段名 image 上传一个 PNG、JPEG 或 WebP 文件。文件最大 10 MiB；宽高各不超过 8192 像素且总像素不超过 4000 万。服务端按文件内容检测格式，不只信任扩展名或 Content-Type。音频、视频、GIF 和 SVG 不被接受。"
          }
        },
        "description": "multipart 图片上传请求。2026-07-24 已在生产以 72,396 字节 PNG 实测成功；字段名必须为 image。"
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "本站 API Key 缺失、无效、已禁用或无法鉴权。入口中间件通常返回扁平的 code/message 格式。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/AnyGatewayErrorResponse"
            },
            "examples": {
              "missingKey": {
                "summary": "缺少 API Key",
                "value": {
                  "code": "API_KEY_REQUIRED",
                  "message": "API key is required in Authorization header (Bearer scheme), x-api-key header, or x-goog-api-key header"
                }
              },
              "invalidKey": {
                "summary": "API Key 无效",
                "value": {
                  "code": "INVALID_API_KEY",
                  "message": "Invalid API key"
                }
              }
            }
          }
        }
      },
...
    }
  }
}